Extract `get_model_files`, `list_model_files_from_argument`; Fix test pollution with throwaway class
